pyproject.toml
[project]
name = "scnrp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "xrpl-py>=4.0.0",
]
